Reworked patching to mainly use patch files

+++ Ishtar/IO/AssetRegistryMerger.cs
@@ -23,15 +23,63 @@ namespace Ishtar.IO
                 Helpers.Log("richTextBox1", $"Could not find AssetRegistry.Json");
                 return;
             }
+            
+            List<string> scanList = new List<string>();
 
-            Parallel.ForEach(infiles, pak =>
+            if (Global.settings.patchMerge)
             {
-                string[] list = Helpers.ListPak(pak);
-                if (list.Any(x => x.Contains("AssetRegistry")))
+                Helpers.Log("richTextBox1", $"Filtering mods with Ishtar Patches...");
+                foreach (string pak in infiles)
                 {
-                    bool added = LoggedPaks.TryAdd(Path.GetFileNameWithoutExtension(pak), pak);
+                    if (!File.Exists(Path.ChangeExtension(pak, ".ishtarPatch")))
+                    {
+                        scanList.Add(pak);
+                    }
+                    else
+                    {
+                        IshtarPatch ishtarPatch = JsonConvert.DeserializeObject<IshtarPatch>(File.ReadAllText(Path.ChangeExtension(pak, ".ishtarPatch")));
+                        try
+                        {
+                            IshtarPatch.Table table = ishtarPatch.tables.First(x => x.Type == IshtarPatch.types.AssetRegistry);
+                            bool added = LoggedPaks.TryAdd(Path.GetFileNameWithoutExtension(pak), pak);
+                        }
+                        catch { }
+                    }
                 }
-            });
+                Helpers.Log("richTextBox1", $"{LoggedPaks.Count} paks filtered");
+            }
+            else
+            {
+                scanList = infiles.ToList();
+            }
+
+            if (Global.settings.scanMerge)
+            {
+                Helpers.Log("richTextBox1", $"Scanning Paks for mergeable files...");
+                if (Global.settings.parallel == true)
+                {
+                    Parallel.ForEach(scanList, pak =>
+                    {
+                        string[] list = Helpers.ListPak(pak);
+                        if (list.Any(x => x.Contains("AssetRegistry")))
+                        {
+                            bool added = LoggedPaks.TryAdd(Path.GetFileNameWithoutExtension(pak), pak);
+                        }
+                    });
+                }
+                else
+                {
+                    foreach (string pak in scanList)
+                    {
+                        string[] list = Helpers.ListPak(pak);
+                        if (list.Any(x => x.Contains("AssetRegistry")))
+                        {
+                            bool added = LoggedPaks.TryAdd(Path.GetFileNameWithoutExtension(pak), pak);
+                        }
+                    }
+                }
+            }
+            
             Helpers.Log("richTextBox1", $"Found {LoggedPaks.Count} paks with Asset Registries for merging");
         }
 
@@ -48,9 +96,7 @@ namespace Ishtar.IO
 
             IEnumerable<string> RegisteredAssetsList = from fassetdata in assestRegistry.fAssetDatas
                                                    select fassetdata.ObjectPath.ToString();
-
             RegisteredAssets = RegisteredAssetsList.ToDictionary(x => x);
-
             Helpers.Log("richTextBox1", $"{RegisteredAssets.ToArray().Length} Base assets registered...");
 
             string outpath = $"ZZZZZ-MergePatch\\CodeVein\\AssetRegistry.bin";
@@ -59,22 +105,31 @@ namespace Ishtar.IO
 
             foreach (string pak in LoggedPaks.Values)
             {
-                string staging = Path.GetFileNameWithoutExtension(pak);
-                if (Directory.Exists(staging))
-                    Directory.Delete(staging, true);
-                if (Path.GetFileName(pak) == "ZZZZZ-MergePatch_P.pak")
-                    continue;
-
-                Helpers.ExtractPak(pak, staging);
-                Helpers.Log("richTextBox1", $"Merging Pak {staging}:");
-
-                string[] list = Directory.GetFiles(staging, "*.*", SearchOption.AllDirectories);
-                string AssRegPath = list.First(x => Path.GetFileName(x) == "AssetRegistry.bin");
-
                 AssetRegistry modRegistry = new AssetRegistry();
-                modRegistry.Read(File.ReadAllBytes(AssRegPath));
+                string staging = Path.GetFileNameWithoutExtension(pak);
 
-                Parallel.ForEach(modRegistry.fAssetDatas, fassetdata =>
+                if (File.Exists(Path.ChangeExtension(pak, ".ishtarPatch")))
+                {
+                    IshtarPatch ishtarPatch = JsonConvert.DeserializeObject<IshtarPatch>(File.ReadAllText(Path.ChangeExtension(pak, ".ishtarPatch")));
+                    IshtarPatch.Table table = ishtarPatch.tables.FirstOrDefault(x => x.Type == IshtarPatch.types.AssetRegistry);
+                    modRegistry = JsonConvert.DeserializeObject<AssetRegistry>(table.DataTable.ToString());
+                }
+                else
+                {
+                    if (Directory.Exists(staging))
+                        Directory.Delete(staging, true);
+                    if (Path.GetFileName(pak) == "ZZZZZ-MergePatch_P.pak")
+                        continue;
+
+                    Helpers.ExtractPak(pak, staging);
+                    string[] list = Directory.GetFiles(staging, "*.*", SearchOption.AllDirectories);
+                    string AssRegPath = list.First(x => Path.GetFileName(x) == "AssetRegistry.bin");
+                    modRegistry.Read(File.ReadAllBytes(AssRegPath));
+                }
+
+                Helpers.Log("richTextBox1", $"Merging Pak {Path.GetFileNameWithoutExtension(pak)}:");
+
+                foreach (FAssetData fassetdata in modRegistry.fAssetDatas)
                 {
                     try
                     {
@@ -86,8 +141,9 @@ namespace Ishtar.IO
                         RegisteredAssets.Add(fassetdata.ObjectPath.ToString(), fassetdata.ObjectPath.ToString());
                         Helpers.Log("richTextBox1", $"    Registered asset {fassetdata.AssetName}");
                     }
-                });
-                Directory.Delete(staging, true);
+                }
+                if (Directory.Exists(staging))
+                    Directory.Delete(staging, true);
             }
 
             Helpers.Log("richTextBox1", $"Writing AssetRegisty to file...");
